data_loader/requirements.txt
azure-storage-blob
azure-identity
httpx
